# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Secure marketplace for trading and auctioning items in Aethoria.\nCommon and Rare items trade via fixed-price listings. Legendary items trade via auctions.",
        "title": "Market Dragon API",
        "contact": {
            "name": "Market Dragon Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/guilds/{id}/wallet": {
            "get": {
                "description": "Return a guild's wallet balance. Available balance is Total minus Reserved.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guilds"
                ],
                "summary": "Get a guild wallet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WalletResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "item not found"
                }
            }
        },
...
        "dto.WalletResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer",
                    "example": 98800
                },
                "guild_id": {
                    "type": "integer",
                    "example": 1
                },
                "reserved": {
                    "type": "integer",
                    "example": 1200
                },
                "total": {
                    "type": "integer",
                    "example": 100000
                }
            }
        }
    },
...
}